Calculate the gradient vector and the Hessian matrix of the weighted log-likelihood function derived from a Cox proportional-hazards model with interval censored data under two-phase sampling design

Calculate the gradient vector and the Hessian matrix of the weighted log-likelihood function derived from a Cox proportional-hazards model with interval censored data under two-phase sampling design

Usage

GENERATE_COX_GRADIENT_HESSIAN(
  Gama,
  d,
  qn,
  ZMat,
  UVec,
  VVec,
  knot,
  degree,
  boundary,
  DELTA,
  OMEGA,
  GENERATE_PHI
)

Arguments

Gama

A vector whose the first d elements are identical with the d-dimensional regrssion coefficients and the last qn elements are identical with the qn-dimensional spline coefficients.

d

The number of unknown coefficients of the regression parameters.

qn

The number of unknown coefficients of the B-spline basis, that is, df or degree + length(knot) + 1 (because the intercept is included).

ZMat

The design matrix corresponding to the regression parameters with the sample size by d.

UVec

A vector consisting of the left observation visit given in the interval-censoring design.

VVec

A vector consisting of the right observation visit given in the interval-censoring design.

knot

The internal breakpoints (i.e., the breakpoints without two ending points) that define the spline. The default is NULL, which results in a basis for the ordinary polynomial regression. Typical values are the mean or median for one knot, quantiles for more knots. See also Boundary.knots.

degree

A non-negative integer degree of the piece-wise polynomial. The default value is 3 for cubic splines. Zero degree is allowed for this function, which is the only difference compared with bs function in package splines. Note that the order of a polynomial equals to the degree plus one.

boundary

A 2-dimensional boundary points at which to anchor the B-spline basis. By default, they are the range of the non-NA data. If both knot and boundary are supplied, the basis parameters do not depend on x. Data can extend beyond boundary.

DELTA

A matrix that each row indicates the interval-censoring type of the observation with the sample size by 3. For example, (1, 0, 0) indicates that UVec is observed but VVec reaches the right boundary, (0, 1, 0) indicates both UVec and VVec are observed and (0, 0, 1) indicates UVec reaches the left boundary but VVec is observed for the observation.

OMEGA

A vector that consisting of the weights for each observation after two-phase sampling. If the two-phase sampling is not used, the values of OMEGA are all 1's.

GENERATE_PHI

The function generates the B-spline basis matrix and the values of a linear combination of B-spline basis function at a given vector.

Value

A list includes: a d + qn dimensional gradient vector evaluated at given Gama, a (d + qn) by (d + qn) Hessian matrix evaluated at given Gama, and a d by d estimated covariance matrix of the regression parameters at given Gama.