Calculate the spline basis matrix and the values of a linear combination of B-spline biasis function at a given vector

Calculate the spline basis matrix and the values of a linear combination of B-spline biasis function at a given vector

Usage

GENERATE_PHI(x, knot, degree, boundary, coeff_spline)

Arguments

x

The predictor vector. Missing values are allowed.

knot

The internal breakpoints (i.e., the breakpoints without two ending points) that define the spline. The default is NULL, which results in a basis for the ordinary polynomial regression. Typical values are the mean or median for one knot, quantiles for more knots. See also Boundary.knots.

degree

A non-negative integer degree of the piece-wise polynomial. The default value is 3 for cubic splines. Zero degree is allowed for this function, which is the only difference compared with bs function in package splines. Note that the order of a polynomial equals to the degree plus one.

boundary

A 2-dimensional boundary points at which to anchor the B-spline basis. By default, they are the range of the non-NA data. If both knot and boundary are supplied, the basis parameters do not depend on x. Data can extend beyond boundary.

coeff_spline

A coefficients vector that are evaluated corresponding to the B-spline basis matrix included an intercept. The length of coeff_spline equals df or degree + length(knot) + 1 (due to the presence of an intercept).

Value

A list includes: a matrix of dimension length(x) by df = degree + length(knot) (plus one if the intercept is included) and a vector that the product of this matrix and the vector coeff_spline. Attributes that correspond to the arguments specified are returned for usage of other functions in this package.

Examples

library(VarEst2Phase)
GENERATE_PHI(x = c(0.1, 0.2, 0.3),
          knot = c(0.25, 0.75),
        degree = 2,
      boundary = c(0, 1),
  coeff_spline = c(0.1, 0.1, 0.1, 0.1, 0.1))
#> $phi_vector
#> [1] 0.1 0.1 0.1
#> 
#> $phi_matrix
#>         1         2          3           4 5
#> [1,] 0.36 0.5866667 0.05333333 0.000000000 0
#> [2,] 0.04 0.7466667 0.21333333 0.000000000 0
#> [3,] 0.00 0.5400000 0.45333333 0.006666667 0
#>